everything neurophotometrics related moved to neurophotometrics, downward compatibility maintained by deprecation warnings tests adjusted, notebooks updated
