+++ tests/test_extractors.py
@@ -5,6 +5,7 @@ import os
 from one.alf.exceptions import ALFObjectNotFound
 
 RUN_EXTRACTOR_TESTS = True if os.environ.get('RUN_EXTRACTOR_TESTS') == '1' else False
+RUN_EXTRACTOR_TESTS = True
 
 
 class PhotometryExtractorTest(unittest.TestCase):
