+++ tests/test_loaders.py
@@ -1,4 +1,4 @@
-import iblphotometry.fpio as fpio
+from iblphotometry import fpio, neurophotometrics
 from tests.base_tests import PhotometryDataTestCase
 import pandas as pd
 
@@ -8,7 +8,7 @@ class TestLoaders(PhotometryDataTestCase):
         versions = ['version_1', 'version_2', 'version_5']  # ...
         for version in versions:
             path = self.versions_path / version / '_neurophotometrics_fpData.raw.pqt'
-            df = fpio.read_neurophotometrics_file(path)
+            df = neurophotometrics.read_neurophotometrics_file(path)
             version_inferred = fpio.infer_neurophotometrics_version_from_data(df)
             assert version == version_inferred
 
@@ -24,16 +24,16 @@ class TestLoaders(PhotometryDataTestCase):
             path = self.versions_path / version / '_neurophotometrics_fpData.digitalInputs.pqt'
             channel = 0 if version == 'version_1' or version == 'version_2' else None
             timestamps_colname = 'Timestamp' if version == 'version_2' else None
-            fpio.read_digital_inputs_file(path, channel=channel, timestamps_colname=timestamps_colname)
+            neurophotometrics.read_digital_inputs_file(path, channel=channel, timestamps_colname=timestamps_colname)
 
     def test_read_neurophotometrics_file(self):
         versions = ['version_1', 'version_2', 'version_5']  # ...
         for version in versions:
             path = self.versions_path / version / '_neurophotometrics_fpData.raw.pqt'
             raw_df = fpio.read_neurophotometrics_file(path)
-            fpio.from_neurophotometrics_df_to_photometry_df(raw_df)
+            neurophotometrics.from_neurophotometrics_df_to_photometry_df(raw_df)
             # the chained version
-            fpio.from_neurophotometrics_file(path)
+            neurophotometrics.from_neurophotometrics_file(path)
 
     # def test_PhotometrySessionLoader(self):
     #     for version in ['version_1', 'version_2', 'version_5']:
